+++ NEWS.md
@@ -9,7 +9,8 @@
    
 ## Minor changes
    
-   
+- The results of `cohens_d()` is now a data frame containing the Cohen's d and the magnitude.
+
 ## Bug fixes
 
 # rstatix 0.1.0